Some clang-tidy aftermath for cmc.
There are lots of code paths in check_mk which are only visible in cmc. We
should really change to git submodules plus subclassing...  :-/

+++ livestatus/src/TableLog.cc
@@ -228,7 +228,7 @@ bool TableLog::isAuthorized(contact *ctc, void *data) {
     host *hst = entry->_host;
 
     if ((hst != nullptr) || (svc != nullptr)) {
-        return is_authorized_for(ctc, hst, svc) != 0;
+        return static_cast<int>(is_authorized_for(ctc, hst, svc)) != 0;
         // suppress entries for messages that belong to
         // hosts that do not exist anymore.
     }
